Move contrast ratio to TextChunk
