+++ src/main/java/org/verapdf/wcag/algorithms/entities/geometry/BoundingBox.java
@@ -10,7 +10,6 @@ public class BoundingBox {
     private double bottomY;
     private double rightX;
     private double topY;
-    private double contrastRatio;
 
     private static double EPSILON = 1.e-18;
 
@@ -122,14 +121,6 @@ public class BoundingBox {
 
     public double getHeight() {
         return (pageNumber.equals(lastPageNumber) && topY > bottomY) ? topY - bottomY : 0;
-    }
-
-    public double getContrastRatio() {
-        return contrastRatio;
-    }
-
-    public void setContrastRatio(double contrastRatio) {
-        this.contrastRatio = contrastRatio;
     }
 
     @Override
